+++ src/main/java/park/brothers/runwith_back/domain/Login/controller/LoginController.java
@@ -1,4 +1,4 @@
-package park.brothers.runwith_back.domain.Login;
+package park.brothers.runwith_back.domain.Login.controller;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
@@ -11,7 +11,9 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import park.brothers.runwith_back.domain.Runner.Runner;
+import park.brothers.runwith_back.domain.Login.dto.LoginRequestDto;
+import park.brothers.runwith_back.domain.Login.service.LoginService;
+import park.brothers.runwith_back.domain.Runner.entity.Runner;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -25,7 +27,7 @@ public class LoginController {
 
     //login api
     @PostMapping("/api/v1/runners/login")
-    public ResponseEntity<Object> login(@RequestBody @Valid LoginForm form,
+    public ResponseEntity<Object> login(@RequestBody @Valid LoginRequestDto form,
                                         BindingResult bindingResult,
                                         HttpServletResponse response) {
